pyproject.toml
[project]
name = "e-sim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "fastapi-users[oauth,sqlalchemy]>=15.0.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
